Project Structure updated
