+++ src/test/java/talrise/runners/CukesRunner.java
@@ -12,7 +12,7 @@ import org.junit.runner.RunWith;
         features = "src/test/resources/features",
         glue = "talrise/step_definitions",
         dryRun = false,
-        tags = "@applications"
+        tags = "@profileSectionPage"
 )
 public class CukesRunner {
 }
