+++ src/test/java/talrise/step_definitions/candidate/normalProfile/ProfileStepDefs.java
@@ -1,4 +1,4 @@
-package talrise.step_definitions.candidate;
+package talrise.step_definitions.candidate.normalProfile;
 
 import io.cucumber.java.en.And;
 import io.cucumber.java.en.Given;
@@ -7,9 +7,6 @@ import io.cucumber.java.en.When;
 import org.junit.Assert;
 import org.openqa.selenium.By;
 import org.openqa.selenium.WebElement;
-import talrise.pages.candidate.LoginPage;
-import talrise.pages.candidate.ProfilePage;
-import talrise.pages.candidate.RegisterPage;
 import talrise.utilities.CommonSteps;
 import talrise.utilities.ConfigurationReader;
