+++ src/test/resources/configuration.properties
@@ -1,6 +1,7 @@
 browser=chrome
 url=http://20.77.188.234/
 environmentUrl=http://20.108.75.9/
+firstLoginUrl=http://20.108.75.9/candidate/personal-info
 adminEmail=ahmet1234@yopmail.com
 password=Test123456!
 superadminEmail=supertest@yopmail.com
