+++ src/test/resources/features/candidate/firstLoginProfile/firstLogin.feature
@@ -1,0 +1,10 @@
+@profileSectionPage
+Feature: Personal Info
+  Scenario: User should be able to see "Complete My Profile Now" after first login
+    Given user goes to Home Page
+#    When logs in successfully
+#    And chooses Candidate as current role
+#    And clicks on "Next" button
+#    Then verifies 'Complete My Profile Now' button is visible
+#    And clicks on 'Complete My Profile Now' button
+#    Then verifies landing on Personal Info page
